Add AppDomain.CurrentDomain.BaseDirectory to assembly search path
NUnit creates a lot of directories for the assemblies, meaning it's hard for the ApiGenerator to find dependencies. By adding *AppDomain.CurrentDomain.BaseDirectory* as search directory, it will have the *actual output directory* of the test project as search directory making sure that dependencies can be found when running using NUnit.